PLN-API/pythonProject4/search-engine/requirements.txt
requests==2.31.0
customtkinter==5.2.2
nltk==3.8.1
scikit-learn==1.4.2
flask==3.0.3
Flask-Cors==4.0.0
scipy==1.10.0
#tensorflow-hub==0.16.1
#tensorflow-text==2.10.0
#tensorflow==2.10.1
transformers==4.41.0
torch==2.3.0
gensim==4.3.2
pyspellchecker==0.8.1
contractions==0.1.73
top2vec==1.0.34
numpy~=1.26.4
bertopic==0.16.2
